PR Feedback: Move to free energy
