+++ timemachine/md/hrex.py
@@ -355,24 +355,6 @@ def get_samples_by_iter_by_replica(
     return samples_by_iter_by_replica
 
 
-def trajectories_by_replica_to_by_state(
-    trajectory_by_iter_by_replica: NDArray,
-    replica_idx_by_state_by_iter: Sequence[Sequence[ReplicaIdx]],
-) -> NDArray:
-    """Utility function to convert the output of `extract_trajectories_by_replica` from (replica, iters, ...) to
-    (state, iters, ...). This is useful for evaluating the trajectories of states.
-    """
-    assert len(trajectory_by_iter_by_replica.shape) == 4
-    replica_idx_by_iter_by_state = np.asarray(replica_idx_by_state_by_iter).T
-    assert replica_idx_by_iter_by_state.shape == trajectory_by_iter_by_replica.shape[:2]
-
-    samples_by_iter_by_state = np.take_along_axis(
-        trajectory_by_iter_by_replica, replica_idx_by_iter_by_state[:, :, None, None], axis=0
-    )
-
-    return samples_by_iter_by_state
-
-
 @dataclass
 class HREXDiagnostics:
     replica_idx_by_state_by_iter: List[List[ReplicaIdx]]
